Fix to Saga artifact bookmark
Updated `amark.png` in both "M15 Saga" and "Mainframe DFC" styles to be more in line with actual existing Artifact Sagas. (Well, the one.)

Substantially lightened the center of the banner to be similar in lightness to the artifact text box background, and updated the very top of the banner to be gold instead of black.