Attempt at fixing ticket, tap and untap symbols
Not sure I understand the situation correctly, but here is what I gather:

Having added `[TK]` to the `mana_sort` allows this function to return `T`, which in turn allows the `mana_filter` function to return `T` as well. Changing `[TK]` to `compound(TK)` seems to work (but TK can no longer be written manually, it must be pasted or inserted wholesale via the insert symbol menu, since `mana_sort` won't allow T or K individually).

The mana_filter_t function is implicitly passed an `in_context: mana_context` argument when called within the `text_filter` function. This was preventing it from returning a T or Q. I believe in all cases it's correct to override that to the empty string.

So `tap_reduction(tap_filter())` was not letting through T when it should, and `mana_filter()` was letting T through when it shouldn't. These two wrongs together made a right where it allowed `T:`, but it did not work in the case of `Q:` since "TK" only contains the letter T and not the letter Q.